+++ SpotlightSample/SpotlightSample/Classes/ConfigViewController.swift
@@ -64,11 +64,12 @@ class ConfigViewController: UITableViewController {
         // ユーザーが選択した設定をUIコントロールから取得します。
         let videoCodec: VideoCodec
         switch videoCodecSegmentedControl.selectedSegmentIndex {
-        case 0: videoCodec = .vp8
-        case 1: videoCodec = .vp9
-        case 2: videoCodec = .av1
-        case 3: videoCodec = .h264
-        case 4: videoCodec = .h265
+        case 0: videoCodec = .default
+        case 1: videoCodec = .vp8
+        case 2: videoCodec = .vp9
+        case 3: videoCodec = .av1
+        case 4: videoCodec = .h264
+        case 5: videoCodec = .h265
         default: fatalError()
         }