+++ CHANGES.md
@@ -23,6 +23,8 @@
   - @zztkm
 - [ADD] 映像コーデックタイプに H.265 を追加する
   - @zztkm
+- [ADD] none 項目がなかった `SimulcastSample` と `SpotlightSample` に none を追加
+  - @zztkm
 
 ## sora-ios-sdk-2024.3.0